requirements.txt
PyQt6>=6.5
PyMuPDF>=1.23
Pillow>=10.0
requests>=2.31
telethon>=1.34
